backend/requirements.txt
fastapi
uvicorn[standard]
geopy
python-dotenv
firecrawl
